#48 refactor: redis 키 생성 로직 분리
