+++ src/main/java/org/sopt/solply_server/global/cache/CachePrefix.java
@@ -1,11 +1,11 @@
-package org.sopt.solply_server.domain.place.constant;
+package org.sopt.solply_server.global.cache;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public enum CachePrefix {
 
-    BOOKMARK_KEY_PREFIX("bookmark");  // 개별 북마크 조회용 키
+    BOOKMARK("bookmark");  // 개별 북마크 조회용 키
 
     private final String prefix;
